rename XdsLoadReportStore to XdsLoadStatsStore
